session-4/pyproject.toml
[project]
name = "lightning-template-hydra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=24.8.0",
    "lightning[extra]>=2.4.0",
    "loguru>=0.7.2",
    "rich>=13.8.1",
    "tensorboard>=2.17.1",
    "timm>=1.0.9",
    "torch>=2.4.1",
    "torchvision>=0.19.1",
    "python-dotenv",
    "scikit-learn",
    "gdown"
]
[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
